% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_dendrogram.R
\name{arrange_dendrogram}
\alias{arrange_dendrogram}
\title{Arrange dendrograms with heatmap}
\usage{
arrange_dendrogram(
  heatmap,
  dendrogram_top,
  dendrogram_right,
  top_right = NULL,
  widths = c(0.8, 0.2),
  heights = c(0.2, 0.8)
)
}
\arguments{
\item{heatmap}{object containing heatmap plot}

\item{dendrogram_top}{top dendrogram plot generated with 
\code{\link{plot_dendrogram}}}

\item{dendrogram_right}{right dendrogram plot generated with 
\code{\link{plot_dendrogram}}}

\item{top_right}{element placed in the top right part of the plot. 
By default \code{NULL} leaving empty space.}

\item{widths}{\code{numeric} vector of heatmap and right dendrogram widths}

\item{heights}{\code{numeric} vector of top dendrogram and heatmap heights}
}
\description{
Arranges both dendrograms with heatmap.
}
\examples{
data(example_data)
dendro_top <- plot_dendrogram(d1)
dendro_right <- plot_dendrogram(d2) + coord_flip()
arrange_dendrogram(hm, dendro_top, dendro_right)
}
