% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Get legend}
\usage{
get_legend(gg_plot)
}
\arguments{
\item{gg_plot}{\code{ggplot} object from which legend will be extracted}
}
\value{
object containing legend
}
\description{
Extracts legend grob from a ggplot object
}
\examples{
data(examples_data)
get_legend(hm)
}
