% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot dendrogram}
\usage{
plot_dendrogram(dendro)
}
\arguments{
\item{dendro}{object of class \code{dendrogram}}
}
\value{
dendrogram plot
}
\description{
Plots dendrogram as a segment with leaf labels.
}
\examples{
data(example_data)
plot_dendrogram(d1)
}
